20171221-qbox-0.15.3(52): 1. 修复NumericConverter类hexDataStringToBytes方法抛出异常的问题

+++ qbox/src/main/java/com/cjq/tool/qbox/util/NumericConverter.java
@@ -129,7 +129,7 @@ public class NumericConverter {
                 if (ss[i].length() > 2) {
                     return null;
                 }
-                data[i] = Byte.parseByte(ss[i], 16);
+                data[i] = (byte) Integer.parseInt(ss[i], 16);
             }
         } catch (NumberFormatException nfe) {
             data = null;
